enh: better code
