+++ src/components/Scaffolding/Scaffolding.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { Component } from 'react';
 import algoliasearch from 'algoliasearch/lite';
 import {
   InstantSearch,
@@ -6,6 +6,8 @@ import {
   SearchBox,
   Configure,
 } from 'react-instantsearch-dom';
+
+import HitBox from './HitBox';
 import './Scaffolding.scss';
 
 const searchClient = algoliasearch(
@@ -13,42 +15,7 @@ const searchClient = algoliasearch(
   '77db8c6516725ef60b8a60223155f7ad'
 );
 
-function htmlDecode (input) {
-  //eslint-disable-next-line
-  return input.replace(/[^a-z0-9\s]/gi, '').replace(/(gt)/gi, '');
-}
-
-const Hit = ({ hit }) => {
-  return (
-    <article className="hit-card">
-      {
-        hit && (
-          <a target="_blank" rel="noopener noreferrer"
-            href={hit.repository && hit.repository.url ? hit.repository.url : `https://npmjs.com/package/${hit.name}`}>
-            <div className="product-wrapper">
-              <div className="product-summary">
-                <div className="product-title">
-                  <h3>{hit.name}</h3> 
-                </div>
-                <div className="product-avatar">
-                  <img src={hit.owner.avatar} height="40" width="40"/>
-                </div>
-              </div>
-              <div className="product-name">
-                <h3>{hit.owner.name}</h3>
-              </div>
-              <div className="product-desc">
-                <p>{ htmlDecode(hit.description) }</p>
-              </div>
-            </div>
-          </a>
-        ) 
-      }
-    </article>
-  );
-};
-
-export default class Scaffolding extends React.Component {
+export default class Scaffolding extends Component {
   render() {
      return (
       <div>
@@ -62,7 +29,7 @@ export default class Scaffolding extends React.Component {
             <SearchBox />
           </div>
           <div className="hits-wrapper">
-              <Hits hitComponent={Hit} />
+              <Hits hitComponent={HitBox} />
           </div>
         </InstantSearch>
       </div>
